src/embed/requirements.txt
biopython >=1.79
tqdm >=2
